src/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
firebird-base==2.0.2
firebird-driver==2.0.2
gunicorn==23.0.0
packaging==25.0
protobuf==5.29.5
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
